Move MTP inference code to separate mixin class
